// MSR.UI/MSR.UI.Answer/src/app/services/mocks/models/createProcedureStepMonitorRequest.ts
/**
 * Parts, Procedures, and Monitors
 * REST API Answer 3
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export class CreateProcedureStepMonitorRequest {
    inputType?: string;
    shouldBe?: string;
    targetValue?: string;
    faultHandling?: string;
    description?: string;
    sendEmailNotification?: boolean;
}